+++ swarm/src/handler/either.rs
@@ -213,6 +213,14 @@ where
                     handler.on_connection_event(ConnectionEvent::AddressChange(address_change))
                 }
             },
+            ConnectionEvent::Datagram(datagram) => match self {
+                Either::Left(handler) => {
+                    handler.on_connection_event(ConnectionEvent::Datagram(datagram))
+                }
+                Either::Right(handler) => {
+                    handler.on_connection_event(ConnectionEvent::Datagram(datagram))
+                }
+            },
             ConnectionEvent::LocalProtocolsChange(supported_protocols) => match self {
                 Either::Left(handler) => handler.on_connection_event(
                     ConnectionEvent::LocalProtocolsChange(supported_protocols),