+++ swarm/src/handler/one_shot.rs
@@ -189,7 +189,8 @@ where
             ConnectionEvent::AddressChange(_)
             | ConnectionEvent::ListenUpgradeError(_)
             | ConnectionEvent::LocalProtocolsChange(_)
-            | ConnectionEvent::RemoteProtocolsChange(_) => {}
+            | ConnectionEvent::RemoteProtocolsChange(_)
+            | ConnectionEvent::Datagram(_) => {}
         }
     }
 }