+++ swarm/src/handler/pending.rs
@@ -88,7 +88,8 @@ impl ConnectionHandler for PendingConnectionHandler {
             | ConnectionEvent::DialUpgradeError(_)
             | ConnectionEvent::ListenUpgradeError(_)
             | ConnectionEvent::LocalProtocolsChange(_)
-            | ConnectionEvent::RemoteProtocolsChange(_) => {}
+            | ConnectionEvent::RemoteProtocolsChange(_)
+            | ConnectionEvent::Datagram(_) => {}
         }
     }
 }