+++ transports/quic/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.15.0
+
+- Support unreliable QUIC datagrams, enabled by default. Tune via `Config::datagram_receive_buffer_size` and `datagram_send_buffer_size`.
+  See [PR XXXX](https://github.com/libp2p/rust-libp2p/pull/XXXX).
+
 ## 0.14.0
 
 - Raise MSRV to 1.88.0.